+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Removes the deprecated `create_list` tracker endpoint function as it is no longer available via API
+
 ## v9.4.1 (2024-08-09)
 
 - Fix `get_next_page_of_children` function for User service with correct filter key